[restpro][pay][items]fix free_num = remain_num - lock_num

+++ shopmanager/flashsale/restpro/views_trade.py
@@ -110,10 +110,13 @@ class ShoppingCartViewSet(viewsets.ModelViewSet):
             raise exceptions.APIException(exc.message)
         except Exception:
             pass
+        if not product.is_onshelf():
+            raise exceptions.APIException(u'商品已下架')
+        
         cart_id = data.get("cart_id", None)
         if cart_id:
-            s_temp = ShoppingCart.objects.filter(item_id=product_id, sku_id=sku_id, status=ShoppingCart.CANCEL,
-                                                 buyer_id=customer.id)
+            s_temp = ShoppingCart.objects.filter(item_id=product_id, sku_id=sku_id, 
+                                                 status=ShoppingCart.CANCEL,buyer_id=customer.id)
             s_temp.delete()
         sku_num = 1
         sku = get_object_or_404(ProductSku, pk=sku_id)
@@ -124,7 +127,7 @@ class ShoppingCartViewSet(viewsets.ModelViewSet):
         
         if not Product.objects.lockQuantity(sku, sku_num):
             raise exceptions.APIException(u'商品库存不足')
-        
+
         if product_id and buyer_id and sku_id:
             shop_cart = ShoppingCart.objects.filter(item_id=product_id, buyer_id=buyer_id, sku_id=sku_id,
                                                     status=ShoppingCart.NORMAL)
@@ -887,7 +890,6 @@ class SaleTradeViewSet(viewsets.ModelViewSet):
                    'default':u'订单不在可支付状态'}
          
         deadline = datetime.datetime.now() - datetime.timedelta(seconds=1500)
-        
         instance = self.get_object()
         if instance.status != SaleTrade.WAIT_BUYER_PAY:
             raise exceptions.APIException(_errmsg.get(instance.status,_errmsg.get('default')))
